+++ crates/bevy_hui/src/styles.rs
@@ -126,6 +126,7 @@ pub struct UiStyleQuery<'w, 's> {
     pub image: Query<'w, 's, &'static mut ImageNode>,
     pub text_fonts: Query<'w, 's, &'static mut TextFont>,
     pub text_colors: Query<'w, 's, &'static mut TextColor>,
+    pub text_shadows: Query<'w, 's, &'static mut TextShadow>,
     pub background: Query<'w, 's, &'static mut BackgroundColor>,
     pub border_radius: Query<'w, 's, &'static mut BorderRadius>,
     pub border_color: Query<'w, 's, &'static mut BorderColor>,
@@ -162,6 +163,13 @@ impl<'w, 's> UiStyleQuery<'w, 's> {
             radius.bottom_right = computed.border_radius.bottom;
             radius.bottom_left = computed.border_radius.left;
         });
+
+        if let Some(computed_shadow) = computed.text_shadow.as_ref() {
+            _ = self.text_shadows.get_mut(entity).map(|mut shadow| {
+                shadow.color = computed_shadow.color;
+                shadow.offset = computed_shadow.offset;
+            });
+        }
 
         if let Some(computed_shadow) = computed.shadow.as_ref() {
             _ = self.shadow.get_mut(entity).map(|mut shadow| {
@@ -302,6 +310,14 @@ impl<'w, 's> UiStyleQuery<'w, 's> {
                     });
                 }
             }
+            StyleAttr::TextShadow(shadow) => {
+                if let Some(computed_shadow) = computed.text_shadow.as_ref() {
+                    _ = self.text_shadows.get_mut(entity).map(|mut s| {
+                        s.offset = computed_shadow.offset.lerp(shadow.offset, ratio);
+                        s.color = lerp_color(&computed_shadow.color, &shadow.color, ratio);
+                    });
+                }
+            }
             StyleAttr::ShadowOffset(x, y) => {
                 if let Some(computed_shadow) = computed.shadow.as_ref() {
                     _ = self.shadow.get_mut(entity).map(|mut shadow| {
@@ -417,6 +433,7 @@ pub struct ComputedStyle {
     pub image_mode: Option<NodeImageMode>,
     pub image_region: Option<Rect>,
     pub shadow: Option<BoxShadow>,
+    pub text_shadow: Option<TextShadow>,
     pub background: Color,
     pub outline: Option<Outline>,
     pub font: Handle<Font>,
@@ -446,6 +463,7 @@ impl Default for ComputedStyle {
             shadow: None,
             image_region: None,
             outline: None,
+            text_shadow: None,
             font: Handle::default(),
             font_size: 12.,
             font_color: Color::WHITE,
@@ -628,6 +646,7 @@ impl HtmlStyle {
                     ));
                 }
             },
+            StyleAttr::TextShadow(shadow) => self.computed.text_shadow = Some(shadow),
             StyleAttr::ShadowOffset(x, y) => match self.computed.shadow.as_mut() {
                 Some(shadow) => {
                     shadow[0].x_offset = x;